factor out waitForCompletedObjectiveId
and directlyFundALedgerChannel
